requirements.txt
openai-whisper
numpy
sounddevice
pystray
pynput
pillow
